split table meta data cache
